Moves Google Analytics service into services directory. Adds search_topic label to GA event for search filter
