+++ frontend/src/app/_services/authentication/authentication-service.ts
@@ -3,7 +3,7 @@ import { HttpClient } from '@angular/common/http';
 import {Observable, BehaviorSubject, Subject, throwError} from 'rxjs';
 import {catchError, map} from 'rxjs/operators';
 import { User } from '../../_models/user';
-import { GoogleAnalyticsService } from '../../google-analytics.service';
+import { GoogleAnalyticsService } from '../google-analytics/google-analytics.service';
 import {StarError} from '../../star-error';
 import {ConfigService} from '../config/config.service';
