+++ frontend/src/app/flow/flow.component.ts
@@ -13,7 +13,7 @@ import {Step, StepStatus} from '../_models/step';
 import {User} from '../_models/user';
 import {ApiService} from '../_services/api/api.service';
 import {AuthenticationService} from '../_services/authentication/authentication-service';
-import {GoogleAnalyticsService} from '../google-analytics.service';
+import {GoogleAnalyticsService} from '../_services/google-analytics/google-analytics.service';
 
 enum FlowState {
   INTRO = 'intro',
