+++ frontend/src/app/login/login.component.ts
@@ -6,7 +6,7 @@ import {DeviceDetectorService} from 'ngx-device-detector';
 import {AuthenticationService} from '../_services/authentication/authentication-service';
 import {scrollToTop} from '../../util/scrollToTop';
 import {User} from '../_models/user';
-import {GoogleAnalyticsService} from '../google-analytics.service';
+import {GoogleAnalyticsService} from '../_services/google-analytics/google-analytics.service';
 
 @Component({
   selector: 'app-login',
