+++ frontend/src/app/password-reset/password-reset.component.ts
@@ -8,7 +8,7 @@ import {PasswordRequirements} from '../_models/password_requirements';
 import {User} from '../_models/user';
 import {ApiService} from '../_services/api/api.service';
 import {AuthenticationService} from '../_services/authentication/authentication-service';
-import {GoogleAnalyticsService} from '../google-analytics.service';
+import {GoogleAnalyticsService} from '../_services/google-analytics/google-analytics.service';
 
 @Component({
   selector: 'app-password-reset',
